backdb/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
greenlet
pydantic>=2.0.0
pydantic-settings
PyJWT==2.8.0
cryptography==42.0.5
bcrypt==4.1.2
python-multipart
python-dotenv
email-validator
pytest
pytest-asyncio
pytest-cov
httpx
azure-keyvault-secrets
azure-identity